requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
google-api-core==2.15.0
pyarrow==14.0.1
protobuf==4.25.3
typing-extensions==4.9.0
python-dotenv==1.0.0
